claude generated file split between drivers and task files
