+++ src/drivers/adcs/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_library(adcs_driver adcs_driver.c)
+
+target_link_libraries(adcs_driver
+    pico_stdlib
+    hardware_uart
+    hardware_irq
+)
+
+target_include_directories(adcs_driver PUBLIC ${CMAKE_CURRENT_LIST_DIR})
